Refactor fastlife_app in tests for security policy on what we want
