+++ tests/fastlife_app/templates/Secured.jinja
@@ -1,5 +1,5 @@
 {# def user #}
 <Layout>
   <H1>Welcome back {{ user.username }}!</H1>
-  <A href="/logout">logout</A>
+  <A href="/admin/logout">logout</A>
 </Layout>
