+++ tests/fastlife_app/views/app/__init__.py
@@ -1,0 +1,1 @@
+"""Dummy app"""
