+++ tests/fastlife_app/views/app/admin/secured.py
