+++ tests/fastlife_app/views/app/forms/booleanfield.py
