+++ tests/fastlife_app/views/app/forms/enumfield.py
