+++ tests/fastlife_app/views/app/forms/enumsfield.py
