+++ tests/fastlife_app/views/app/forms/floatfield.py
