+++ tests/fastlife_app/views/app/forms/string_sequence.py
