+++ tests/fastlife_app/views/app/hello_fstring.py
