+++ tests/fastlife_app/views/app/home.py
