+++ tests/fastlife_app/views/app/i18n/__init__.py
