+++ tests/fastlife_app/views/app/i18n/dummy_messages.py
