+++ tests/fastlife_app/views/app/i18n/hello_i18n.py
