+++ tests/fastlife_app/views/app/icons.py
