+++ tests/unittests/app/conftest.py
@@ -9,9 +9,9 @@ from fastlife.testing import WebTestClient
 @pytest.fixture
 async def app(settings: Settings):
     conf = Configurator(settings=settings)
-    conf.include("tests.fastlife_app.security")
-    conf.include("tests.fastlife_app.views", ignore=".api")
+    conf.include("tests.fastlife_app.views", ignore=[".api", ".app.admin"])
     conf.include("tests.fastlife_app.views.api", route_prefix="/api")
+    conf.include("tests.fastlife_app.views.app.admin", route_prefix="/admin")
     yield conf.build_asgi_app()
